#126 Move images related stuff inside its module
